Backtest/requirements.txt
MetaTrader5==5.0.45
pandas==2.2.0
streamlit==1.31.0
gspread==6.0.0
google-auth==2.27.0
ta==0.11.0
python-dotenv==1.0.1
numpy==1.26.4
requests==2.31.0